pyproject.toml
[tool.poetry]
name = "book-store-api"
version = "0.1.0"
description = ""
authors = ["AbdurashiDIDO <abdurashiddido@gmail.com>"]
readme = "README.md"
packages = [{include = "book_store_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
pypdf2 = "^3.0.1"
pycryptodome = "^3.18.0"
pillow = "^10.0.0"
pdf2image = "^1.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
